+++ CHANGELOG.md
@@ -8,15 +8,18 @@ Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
 ## [Unreleased]
+#### Added
+- Polls
+
 #### Fixed
 - A bug with filter for `paymentMethodType` on pay page
+- Hidden 'Polls' 
 
 ## [1.7.0-rc.0] - 2020-01-24
 #### Added
 - Filter for `paymentMethodType` on pay page
 - Update Promo Code form
 - Update offer form
-- Polls
 
 #### Changed
 - KYC recovery without filling out a form